Huge (parallel, undirected, unweighted) louvain (#550)
* adapt trianglecount to work with >2bn nodes

* adapt testcases

* adapt writeback methods

* set writemillis

* issue 542

* mv specialized translators into their datatype classes

* WIP

* WIP

* refactoring & jdoc

* keep it object

* fix test

* debugging

* fix test

* fix test

+++ algo/src/main/java/org/neo4j/graphalgo/impl/louvain/WeightedLouvain.java
@@ -21,6 +21,7 @@ package org.neo4j.graphalgo.impl.louvain;
 import org.neo4j.graphalgo.api.Graph;
 import org.neo4j.graphalgo.core.utils.ParallelUtil;
 import org.neo4j.graphalgo.core.utils.ProgressLogger;
+import org.neo4j.graphalgo.core.utils.paged.LongArray;
 import org.neo4j.graphalgo.core.utils.traverse.SimpleBitSet;
 import org.neo4j.graphalgo.impl.Algorithm;
 import org.neo4j.graphdb.Direction;
@@ -180,7 +181,7 @@ public class WeightedLouvain extends Algorithm<WeightedLouvain> implements Louva
     }
 
     @Override
-    public int getCommunityCount() {
+    public long getCommunityCount() {
         final SimpleBitSet bitSet = new SimpleBitSet(nodeCount);
         for (int i = 0; i < nodeCount; i++) {
             bitSet.put(nodeCommunity[i]);
